Color plugged in Web
git-svn-id: http://www.geogebra.org/svn/trunk/geogebra@14437 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ desktop/geogebra/factories/AwtFactory.java
@@ -33,7 +33,7 @@ public class AwtFactory extends geogebra.common.factories.AwtFactory{
 	}
 
 	@Override
-	public Color newColor(int red, float green, int blue) {
+	public Color newColor(float red, float green, float blue) {
 		return new geogebra.awt.Color(red, green, blue);
 	}
